Rename module to traces.py
